requirements.txt
requests
pyyaml
schedule
python-systemd